requirements.txt
streamlit==1.31.1
boto3==1.34.34
Pillow==10.2.0
python-magic==0.4.27
python-magic-bin==0.4.14